ml/word/requirements.txt
# Word CRNN training
tensorflow>=2.15
tensorflowjs>=4.0
numpy>=1.24
Pillow>=10.0
